# api/api-docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgDx API",
    "description": "AgDx API facilitates sharing of crop pest and disease data between platforms and/or by providing links between complementary ICT tools where appropriate.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0.0-beta1"
  },
...
  "servers": [
    {
      "url": "http://localhost:9100",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/crop/update/{id}": {
      "put": {
        "tags": [
          "Crop data"
        ],
        "summary": "Update crop record",
        "operationId": "updateCropData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
...
    "/api/v1/crop/add": {
      "post": {
        "tags": [
          "Crop data"
        ],
        "summary": "Add new crop",
        "operationId": "addCropData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
...
    "/api/v1/crop/{cropName}/scientific": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using scientific name",
        "operationId": "getCropByScientificName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/{cropName}/common": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using common name",
        "operationId": "getCropByCommonName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crop/ontology/{ontologyId}": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using ontology id",
        "operationId": "getCropByOntologyId",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/genus/{genus}": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using genus",
        "operationId": "getCropByGenus",
        "parameters": [
          {
            "name": "genus",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/family/{family}": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using family",
        "operationId": "getCropByFamily",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CropEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "commonName": {
            "type": "string",
            "description": "Crop common name",
            "example": "Banana"
          },
          "scientificName": {
            "type": "string",
            "description": "Crop scientific name",
            "example": "Musca"
          },
          "ontologyId": {
            "type": "string",
            "description": "Ontology id of the crop",
            "example": "C350"
          },
          "genus": {
            "type": "string",
            "description": "Genus crop",
            "example": "Musa"
          },
          "family": {
            "type": "string",
            "description": "Family crop",
            "example": "Musaceae"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Sort": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageCropEntity": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CropEntity"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/Pageable"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api": {
        "type": "http",
        "in": "header",
        "scheme": "basic"
      }
    }
  }
}